@@ -3,20 +3,20 @@
     "start": "vite --open"
   },
   "dependencies": {
-    "@faker-js/faker": "^6.3.1",
-    "@tanstack/react-table": "8.0.0-beta.9",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "typescript": "^4.7.4"
+    "@faker-js/faker": "^8.1.0",
+    "@tanstack/react-table": "8.10.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "typescript": "^5.2.2"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "4.0.0",
-    "@tanstack/table-core": "^8.0.0",
-    "@types/faker": "^5.5.9",
-    "@types/react": "^18.0.14",
-    "@types/react-dom": "^18.0.5",
-    "@vitejs/plugin-react": "^1.3.2",
-    "rollup": "^2.75.7",
-    "vite": "2.8.6"
+    "@rollup/plugin-replace": "5.0.2",
+    "@tanstack/table-core": "^8.10.1",
+    "@types/faker": "^6.6.9",
+    "@types/react": "^18.2.22",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react": "^4.1.0",
+    "rollup": "^3.29.3",
+    "vite": "4.4.9"
   }
 }
