@@ -1,13 +1,14 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 
 import './index.css'
 
 import {
-  createTable,
-  useTableInstance,
+  useReactTable,
   ColumnResizeMode,
   getCoreRowModel,
+  ColumnDef,
+  flexRender,
 } from '@tanstack/react-table'
 
 type Person = {
@@ -46,52 +47,56 @@ const defaultData: Person[] = [
   },
 ]
 
-let table = createTable().setRowType<Person>()
-
-const defaultColumns = [
-  table.createGroup({
+const defaultColumns: ColumnDef<Person>[] = [
+  {
     header: 'Name',
     footer: props => props.column.id,
     columns: [
-      table.createDataColumn('firstName', {
+      {
+        accessorKey: 'firstName',
         cell: info => info.getValue(),
         footer: props => props.column.id,
-      }),
-      table.createDataColumn(row => row.lastName, {
+      },
+      {
+        accessorFn: row => row.lastName,
         id: 'lastName',
         cell: info => info.getValue(),
         header: () => <span>Last Name</span>,
         footer: props => props.column.id,
-      }),
+      },
     ],
-  }),
-  table.createGroup({
+  },
+  {
     header: 'Info',
     footer: props => props.column.id,
     columns: [
-      table.createDataColumn('age', {
+      {
+        accessorKey: 'age',
         header: () => 'Age',
         footer: props => props.column.id,
-      }),
-      table.createGroup({
+      },
+      {
         header: 'More Info',
         columns: [
-          table.createDataColumn('visits', {
+          {
+            accessorKey: 'visits',
             header: () => <span>Visits</span>,
             footer: props => props.column.id,
-          }),
-          table.createDataColumn('status', {
+          },
+          {
+            accessorKey: 'status',
             header: 'Status',
             footer: props => props.column.id,
-          }),
-          table.createDataColumn('progress', {
+          },
+          {
+            accessorKey: 'progress',
             header: 'Profile Progress',
             footer: props => props.column.id,
-          }),
+          },
         ],
-      }),
+      },
     ],
-  }),
+  },
 ]
 
 function App() {
@@ -101,11 +106,11 @@ function App() {
   ])
 
   const [columnResizeMode, setColumnResizeMode] =
-    React.useState<ColumnResizeMode>('onChange')
+      React.useState<ColumnResizeMode>('onChange')
 
   const rerender = React.useReducer(() => ({}), {})[1]
 
-  const instance = useTableInstance(table, {
+  const table = useReactTable({
     data,
     columns,
     columnResizeMode,
@@ -116,282 +121,296 @@ function App() {
   })
 
   return (
-    <div className="p-2">
-      <select
-        value={columnResizeMode}
-        onChange={e => setColumnResizeMode(e.target.value as ColumnResizeMode)}
-        className="border p-2 border-black rounded"
-      >
-        <option value="onEnd">Resize: "onEnd"</option>
-        <option value="onChange">Resize: "onChange"</option>
-      </select>
-      <div className="h-4" />
-      <div className="text-xl">{'<table/>'}</div>
-      <div className="overflow-x-auto">
-        <table
-          {...{
-            style: {
-              width: instance.getCenterTotalSize(),
-            },
-          }}
+      <div className="p-2">
+        <select
+            value={columnResizeMode}
+            onChange={e => setColumnResizeMode(e.target.value as ColumnResizeMode)}
+            className="border p-2 border-black rounded"
         >
-          <thead>
-          {instance.getHeaderGroups().map(headerGroup => (
-            <tr key={headerGroup.id}>
-              {headerGroup.headers.map(header => (
-                <th
-                  {...{
-                    key: header.id,
-                    colSpan: header.colSpan,
-                    style: {
-                      width: header.getSize(),
-                    },
-                  }}
-                >
-                  {header.isPlaceholder ? null : header.renderHeader()}
-                  <div
-                    {...{
-                      onMouseDown: header.getResizeHandler(),
-                      onTouchStart: header.getResizeHandler(),
-                      className: `resizer ${
-                        header.column.getIsResizing() ? 'isResizing' : ''
-                      }`,
-                      style: {
-                        transform:
-                          columnResizeMode === 'onEnd' &&
-                          header.column.getIsResizing()
-                            ? `translateX(${
-                              instance.getState().columnSizingInfo
-                                .deltaOffset
-                            }px)`
-                            : '',
-                      },
-                    }}
-                  />
-                </th>
-              ))}
-            </tr>
-          ))}
-          </thead>
-          <tbody>
-          {instance.getRowModel().rows.map(row => (
-            <tr key={row.id}>
-              {row.getVisibleCells().map(cell => (
-                <td
-                  {...{
-                    key: cell.id,
-                    style: {
-                      width: cell.column.getSize(),
-                    },
-                  }}
-                >
-                  {cell.renderCell()}
-                </td>
-              ))}
-            </tr>
-          ))}
-          </tbody>
-        </table>
-      </div>
-      <div className="h-4" />
-      <div className="text-xl">{'<div/> (relative)'}</div>
-      <div className="overflow-x-auto">
-        <div
-          {...{
-            className: 'divTable',
-            style: {
-              width: instance.getTotalSize(),
-            },
-          }}
-        >
-          <div className="thead">
-            {instance.getHeaderGroups().map(headerGroup => (
-              <div
-                {...{
-                  key: headerGroup.id,
-                  className: 'tr',
-                }}
-              >
-                {headerGroup.headers.map(header => (
-                  <div
-                    {...{
-                      key: header.id,
-                      className: 'th',
-                      style: {
-                        width: header.getSize(),
-                      },
-                    }}
-                  >
-                    {header.isPlaceholder ? null : header.renderHeader()}
-                    <div
-                      {...{
-                        onMouseDown: header.getResizeHandler(),
-                        onTouchStart: header.getResizeHandler(),
-                        className: `resizer ${
-                          header.column.getIsResizing() ? 'isResizing' : ''
-                        }`,
-                        style: {
-                          transform:
-                            columnResizeMode === 'onEnd' &&
-                            header.column.getIsResizing()
-                              ? `translateX(${
-                                instance.getState().columnSizingInfo
-                                  .deltaOffset
-                              }px)`
-                              : '',
-                        },
-                      }}
-                    />
-                  </div>
-                ))}
-              </div>
-            ))}
-          </div>
-          <div
-            {...{
-              className: 'tbody',
-            }}
+          <option value="onEnd">Resize: "onEnd"</option>
+          <option value="onChange">Resize: "onChange"</option>
+        </select>
+        <div className="h-4" />
+        <div className="text-xl">{'<table/>'}</div>
+        <div className="overflow-x-auto">
+          <table
+              {...{
+                style: {
+                  width: table.getCenterTotalSize(),
+                },
+              }}
           >
-            {instance.getRowModel().rows.map(row => (
-              <div
-                {...{
-                  key: row.id,
-                  className: 'tr',
-                }}
-              >
-                {row.getVisibleCells().map(cell => (
-                  <div
-                    {...{
-                      key: cell.id,
-                      className: 'td',
-                      style: {
-                        width: cell.column.getSize(),
-                      },
-                    }}
-                  >
-                    {cell.renderCell()}
-                  </div>
-                ))}
-              </div>
+            <thead>
+            {table.getHeaderGroups().map(headerGroup => (
+                <tr key={headerGroup.id}>
+                  {headerGroup.headers.map(header => (
+                      <th
+                          {...{
+                            key: header.id,
+                            colSpan: header.colSpan,
+                            style: {
+                              width: header.getSize(),
+                            },
+                          }}
+                      >
+                        {header.isPlaceholder
+                            ? null
+                            : flexRender(
+                                header.column.columnDef.header,
+                                header.getContext()
+                            )}
+                        <div
+                            {...{
+                              onMouseDown: header.getResizeHandler(),
+                              onTouchStart: header.getResizeHandler(),
+                              className: `resizer ${
+                                  header.column.getIsResizing() ? 'isResizing' : ''
+                              }`,
+                              style: {
+                                transform:
+                                    columnResizeMode === 'onEnd' &&
+                                    header.column.getIsResizing()
+                                        ? `translateX(${
+                                            table.getState().columnSizingInfo.deltaOffset
+                                        }px)`
+                                        : '',
+                              },
+                            }}
+                        />
+                      </th>
+                  ))}
+                </tr>
             ))}
+            </thead>
+            <tbody>
+            {table.getRowModel().rows.map(row => (
+                <tr key={row.id}>
+                  {row.getVisibleCells().map(cell => (
+                      <td
+                          {...{
+                            key: cell.id,
+                            style: {
+                              width: cell.column.getSize(),
+                            },
+                          }}
+                      >
+                        {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                      </td>
+                  ))}
+                </tr>
+            ))}
+            </tbody>
+          </table>
+        </div>
+        <div className="h-4" />
+        <div className="text-xl">{'<div/> (relative)'}</div>
+        <div className="overflow-x-auto">
+          <div
+              {...{
+                className: 'divTable',
+                style: {
+                  width: table.getTotalSize(),
+                },
+              }}
+          >
+            <div className="thead">
+              {table.getHeaderGroups().map(headerGroup => (
+                  <div
+                      {...{
+                        key: headerGroup.id,
+                        className: 'tr',
+                      }}
+                  >
+                    {headerGroup.headers.map(header => (
+                        <div
+                            {...{
+                              key: header.id,
+                              className: 'th',
+                              style: {
+                                width: header.getSize(),
+                              },
+                            }}
+                        >
+                          {header.isPlaceholder
+                              ? null
+                              : flexRender(
+                                  header.column.columnDef.header,
+                                  header.getContext()
+                              )}
+                          <div
+                              {...{
+                                onMouseDown: header.getResizeHandler(),
+                                onTouchStart: header.getResizeHandler(),
+                                className: `resizer ${
+                                    header.column.getIsResizing() ? 'isResizing' : ''
+                                }`,
+                                style: {
+                                  transform:
+                                      columnResizeMode === 'onEnd' &&
+                                      header.column.getIsResizing()
+                                          ? `translateX(${
+                                              table.getState().columnSizingInfo.deltaOffset
+                                          }px)`
+                                          : '',
+                                },
+                              }}
+                          />
+                        </div>
+                    ))}
+                  </div>
+              ))}
+            </div>
+            <div
+                {...{
+                  className: 'tbody',
+                }}
+            >
+              {table.getRowModel().rows.map(row => (
+                  <div
+                      {...{
+                        key: row.id,
+                        className: 'tr',
+                      }}
+                  >
+                    {row.getVisibleCells().map(cell => (
+                        <div
+                            {...{
+                              key: cell.id,
+                              className: 'td',
+                              style: {
+                                width: cell.column.getSize(),
+                              },
+                            }}
+                        >
+                          {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                        </div>
+                    ))}
+                  </div>
+              ))}
+            </div>
           </div>
         </div>
-      </div>
-      <div className="h-4" />
-      <div className="text-xl">{'<div/> (absolute positioning)'}</div>
-      <div className="overflow-x-auto">
-        <div
-          {...{
-            className: 'divTable',
-            style: {
-              width: instance.getTotalSize(),
-            },
-          }}
-        >
-          <div className="thead">
-            {instance.getHeaderGroups().map(headerGroup => (
-              <div
-                {...{
-                  key: headerGroup.id,
-                  className: 'tr',
-                  style: {
-                    position: 'relative',
-                  },
-                }}
-              >
-                {headerGroup.headers.map(header => (
+        <div className="h-4" />
+        <div className="text-xl">{'<div/> (absolute positioning)'}</div>
+        <div className="overflow-x-auto">
+          <div
+              {...{
+                className: 'divTable',
+                style: {
+                  width: table.getTotalSize(),
+                },
+              }}
+          >
+            <div className="thead">
+              {table.getHeaderGroups().map(headerGroup => (
                   <div
-                    {...{
-                      key: header.id,
-                      className: 'th',
-                      style: {
-                        position: 'absolute',
-                        left: header.getStart(),
-                        width: header.getSize(),
-                      },
-                    }}
-                  >
-                    {header.isPlaceholder ? null : header.renderHeader()}
-                    <div
                       {...{
-                        onMouseDown: header.getResizeHandler(),
-                        onTouchStart: header.getResizeHandler(),
-                        className: `resizer ${
-                          header.column.getIsResizing() ? 'isResizing' : ''
-                        }`,
+                        key: headerGroup.id,
+                        className: 'tr',
                         style: {
-                          transform:
-                            columnResizeMode === 'onEnd' &&
-                            header.column.getIsResizing()
-                              ? `translateX(${
-                                instance.getState().columnSizingInfo
-                                  .deltaOffset
-                              }px)`
-                              : '',
+                          position: 'relative',
                         },
                       }}
-                    />
-                  </div>
-                ))}
-              </div>
-            ))}
-          </div>
-          <div
-            {...{
-              className: 'tbody',
-            }}
-          >
-            {instance.getRowModel().rows.map(row => (
-              <div
-                {...{
-                  key: row.id,
-                  className: 'tr',
-                  style: {
-                    position: 'relative',
-                  },
-                }}
-              >
-                {row.getVisibleCells().map(cell => (
-                  <div
-                    {...{
-                      key: cell.id,
-                      className: 'td',
-                      style: {
-                        position: 'absolute',
-                        left: cell.column.getStart(),
-                        width: cell.column.getSize(),
-                      },
-                    }}
                   >
-                    {cell.renderCell()}
+                    {headerGroup.headers.map(header => (
+                        <div
+                            {...{
+                              key: header.id,
+                              className: 'th',
+                              style: {
+                                position: 'absolute',
+                                left: header.getStart(),
+                                width: header.getSize(),
+                              },
+                            }}
+                        >
+                          {header.isPlaceholder
+                              ? null
+                              : flexRender(
+                                  header.column.columnDef.header,
+                                  header.getContext()
+                              )}
+                          <div
+                              {...{
+                                onMouseDown: header.getResizeHandler(),
+                                onTouchStart: header.getResizeHandler(),
+                                className: `resizer ${
+                                    header.column.getIsResizing() ? 'isResizing' : ''
+                                }`,
+                                style: {
+                                  transform:
+                                      columnResizeMode === 'onEnd' &&
+                                      header.column.getIsResizing()
+                                          ? `translateX(${
+                                              table.getState().columnSizingInfo.deltaOffset
+                                          }px)`
+                                          : '',
+                                },
+                              }}
+                          />
+                        </div>
+                    ))}
                   </div>
-                ))}
-              </div>
-            ))}
+              ))}
+            </div>
+            <div
+                {...{
+                  className: 'tbody',
+                }}
+            >
+              {table.getRowModel().rows.map(row => (
+                  <div
+                      {...{
+                        key: row.id,
+                        className: 'tr',
+                        style: {
+                          position: 'relative',
+                        },
+                      }}
+                  >
+                    {row.getVisibleCells().map(cell => (
+                        <div
+                            {...{
+                              key: cell.id,
+                              className: 'td',
+                              style: {
+                                position: 'absolute',
+                                left: cell.column.getStart(),
+                                width: cell.column.getSize(),
+                              },
+                            }}
+                        >
+                          {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                        </div>
+                    ))}
+                  </div>
+              ))}
+            </div>
           </div>
         </div>
-      </div>
-      <div className="h-4" />
-      <button onClick={() => rerender()} className="border p-2">
-        Rerender
-      </button>
-      <pre>
+        <div className="h-4" />
+        <button onClick={() => rerender()} className="border p-2">
+          Rerender
+        </button>
+        <pre>
         {JSON.stringify(
-          {
-            columnSizing: instance.getState().columnSizing,
-            columnSizingInfo: instance.getState().columnSizingInfo,
-          },
-          null,
-          2
+            {
+              columnSizing: table.getState().columnSizing,
+              columnSizingInfo: table.getState().columnSizingInfo,
+            },
+            null,
+            2
         )}
       </pre>
-    </div>
+      </div>
   )
 }
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+const rootElement = document.getElementById('root')
+if (!rootElement) throw new Error('Failed to find the root element')
+
+ReactDOM.createRoot(rootElement).render(
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
 )
